Move routes definition to the mainRouter file
