+++ setup/react/app/MainRouter.js
@@ -3,7 +3,6 @@ import { withRouter } from 'react-router';
 import { Route, Switch } from 'react-router-dom';
 import { addSearchParamsToUrl } from 'src/utils/searchParams';
 import ConnectionContext from '@libs/wcm/context/connectionContext';
-import useWalletConnectEventsManager from '@libs/wcm/hooks/useConnectionEventsManager';
 import { EVENTS } from '@libs/wcm/constants/lifeCycle';
 import routesMap from 'src/routes/routesMap';
 import NotFound from 'src/modules/common/components/NotFound';
@@ -13,7 +12,7 @@ import styles from './app.css';
 
 const MainRouter = ({ history }) => {
   const { events } = useContext(ConnectionContext);
-  useWalletConnectEventsManager();
+  const routesList = Object.keys(routes);
 
   useEffect(() => {
     if (events.length && events[events.length - 1].name === EVENTS.SESSION_REQUEST) {
@@ -24,20 +23,18 @@ const MainRouter = ({ history }) => {
   return (
     <div className={`${styles.mainContent} ${styles.mainBox}`}>
       <Switch>
-        {
-          Object.entries(routes).map(([key, route]) => (
-            <CustomRoute
-              key={route.path}
-              route={route}
-              path={route.path}
-              exact={route.exact}
-              isPrivate={route.isPrivate}
-              forbiddenTokens={route.forbiddenTokens}
-              component={routesMap[key]}
-              history={history}
-            />
-          ))
-        }
+        {routesList.map((route) => (
+          <CustomRoute
+            key={routes[route].path}
+            route={routes[route]}
+            path={routes[route].path}
+            exact={routes[route].exact}
+            isPrivate={routes[route].isPrivate}
+            forbiddenTokens={routes[route].forbiddenTokens}
+            component={routesMap[route]}
+            history={history}
+          />
+        ))}
         <Route path="*" component={NotFound} />
       </Switch>
     </div>
