+++ setup/react/app/index.js
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 // This is covered by e2e tests
 import React, { useEffect, useState } from 'react';
-import { Route, Switch, withRouter } from 'react-router-dom';
+import { withRouter } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { settingsRetrieved } from 'src/modules/settings/store/actions';
 import { ToastContainer } from 'react-toastify';
@@ -9,17 +9,14 @@ import 'react-toastify/dist/ReactToastify.css';
 import { hot } from 'react-hot-loader/root';
 import { bookmarksRetrieved } from 'src/modules/bookmark/store/action';
 import { watchListRetrieved } from 'src/modules/pos/validator/store/actions/watchList';
-import NotFound from 'src/modules/common/components/NotFound';
 import useIpc from '@update/hooks/useIpc';
 import ConnectionProvider from '@libs/wcm/context/connectionProvider';
 import FlashMessageHolder from 'src/theme/flashMessage/holder';
 import client from 'src/utils/api/client';
 import DialogHolder from 'src/theme/dialog/holder';
 import OfflineWrapper from 'src/modules/common/components/offlineWrapper';
-import CustomRoute from 'src/modules/common/components/customRoute';
 import NavigationBars from 'src/modules/common/components/bars';
 import ThemeContext from 'src/theme/themeProvider';
-import routesMap from 'src/routes/routesMap';
 import { useTransactionUpdate } from '@transaction/hooks';
 import routes from 'src/routes/routes';
 import { MOCK_SERVICE_WORKER } from 'src/const/config';
@@ -28,6 +25,7 @@ import {
   useApplicationManagement,
   useCurrentApplication,
 } from 'src/modules/blockchainApplication/manage/hooks';
+import MainRouter from './MainRouter';
 import './variables.css';
 import styles from './app.css';
 
@@ -68,7 +66,6 @@ const App = ({ history }) => {
   }, [isLoading, chainMetaData]);
   useTransactionUpdate(loaded);
 
-  const routesList = Object.keys(routes);
   const routeObj = Object.values(routes).find((r) => r.path === history.location.pathname) || {};
   return (
     <ConnectionProvider>
@@ -93,23 +90,7 @@ const App = ({ history }) => {
           <main className={`${styles.bodyWrapper} ${loaded ? styles.loaded : ''}`}>
             <section className="scrollContainer">
               <FlashMessageHolder />
-              <div className={`${styles.mainContent} ${styles.mainBox}`}>
-                <Switch>
-                  {routesList.map((route) => (
-                    <CustomRoute
-                      key={routes[route].path}
-                      route={routes[route]}
-                      path={routes[route].path}
-                      exact={routes[route].exact}
-                      isPrivate={routes[route].isPrivate}
-                      forbiddenTokens={routes[route].forbiddenTokens}
-                      component={routesMap[route]}
-                      history={history}
-                    />
-                  ))}
-                  <Route path="*" component={NotFound} />
-                </Switch>
-              </div>
+              <MainRouter />
             </section>
           </main>
         </OfflineWrapper>
