+++ src/routes/routes.js
@@ -271,7 +271,7 @@ export const modals = {
     isPrivate: false,
     forbiddenTokens: [],
   },
-  RequestView: {
+  requestView: {
     isPrivate: false,
     forbiddenTokens: [],
   },
